requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
defusedxml>=0.7.1
ttkbootstrap>=1.10.0
tkinterdnd2>=0.3.0
pytest>=7.0.0
